20250427/python/pyproject.toml
[project]
name = "rtsp-demo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
]
